Version 2025.12.8
更新内容：
1.添加了许多图片素材，用于制作功能菜单系统。
2.更新了NPC的战斗逻辑，由之前的必须能够释放法术才会移动，改为了若无法释放任何一个法术，则会先朝目标接近。

修复部分：
1.修复了境界系统的数据传输问题，如无法读取到特定境界的数据。
2.修复了战斗过程中反应值累计的速率问题。

+++ Assets/Scripts/UI/Combat/CombatUI.cs
@@ -57,7 +57,7 @@ namespace TXDCL.Combat
             DaoCangPanelUI.Instance.InitializeDaoCangPanel(character);
         }
 
-        public void UpdateCharactersTurnProgressUI(CharacterBase character, int value)
+        public void UpdateCharactersTurnProgressUI(CharacterBase character, float value)
         {
             if(!activeCharacters.TryGetValue(character, out var activeCharacter)) return;
             activeCharacter.transform.localPosition = new Vector3(value < 0 ? Mathf.Max(-1000, -400 + value * 0.4f) : Mathf.Min(1000, -400 + value * 0.4f), 10, 0);
